server/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
sse-starlette>=2.1.0
python-dotenv>=1.0.0
crewai[google-genai]>=0.80.0
fpdf2>=2.7.0
